requirements.txt
Django==5.1.6
djangorestframework==3.15.2
mysqlclient==2.2.7
requests==2.32.3